chore(vmix): refactor: move VMixInputHandler to separate file
